requirements.txt
# Multi-Camera Gimbal System Requirements
# Raspberry Pi 5 - Ubuntu 22.04/24.04

# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
picamera2>=0.3.12

# GPIO and Hardware Control
RPi.GPIO>=0.7.1
gpiozero>=2.0
lgpio>=0.2.0
pigpio>=1.78

# I2C and SPI for sensors
smbus2>=0.4.2
spidev>=3.6
adafruit-circuitpython-bh1750>=1.1.0
adafruit-circuitpython-pca9685>=3.4.0

# Serial communication
pyserial>=3.5

# Video processing and encoding
ffmpeg-python>=0.2.0
av>=10.0.0

# Network and streaming
flask>=3.0.0
socketio>=5.9.0
python-socketio>=5.9.0
websockets>=12.0

# RF and wireless
pyrf24>=1.4.0

# Display and UI
pygame>=2.5.0
PyQt5>=5.15.9
pillow>=10.0.0

# Multi-threading
threading
multiprocessing

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
